# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TrashTrack API",
        "description": "TrashTrack adalah API untuk pelacakan sampah, pelaporan insiden, dan manajemen data sampah pengguna.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v2/trash": {
            "get": {
                "tags": ["Trash V2"],
                "summary": "Dapatkan daftar sampah ringkas",
                "description": "Mengambil daftar sampah dengan informasi yang disederhanakan",
                "operationId": "50517c0e1a1918b376f5b81c0e2a4995",
                "responses": {
                    "200": {
                        "description": "Daftar sampah berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Botol Plastik"
                                            },
                                            "category": {
                                                "type": "string",
                                                "example": "Anorganik"
                                            },
                                            "weight": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 2.5
                                            },
                                            "weight_unit": {
                                                "type": "string",
                                                "example": "kg"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": ["Trash V2"],
                "summary": "Buat data sampah ringkas",
                "description": "Menambahkan data sampah dan mengembalikan format ringkas",
                "operationId": "96a8a4594c5919f79f5ef7588ea3de13",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": ["name", "category", "weight"],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Botol Plastik"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Anorganik"
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data sampah berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Botol Plastik"
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "Anorganik"
                                        },
                                        "weight": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 2.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/trash/{id}": {
            "get": {
                "tags": ["Trash V2"],
                "summary": "Dapatkan detail sampah ringkas",
                "description": "Mengambil detail ringkas sampah berdasarkan ID",
                "operationId": "430687759c3ed409e0ffc9693725f24b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID sampah",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail sampah berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Botol Plastik"
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "Anorganik"
                                        },
                                        "weight": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 2.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid"
                    },
                    "404": {
                        "description": "Sampah tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/trash": {
            "get": {
                "tags": ["Trash"],
                "summary": "Dapatkan daftar sampah",
                "description": "Mengambil seluruh data sampah yang telah dicatat",
                "operationId": "ddffcb2862d79f04ef04e0b67d9d5f83",
                "responses": {
                    "200": {
                        "description": "Daftar sampah berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Trash"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": ["Trash"],
                "summary": "Buat data sampah baru",
                "description": "Menambahkan data sampah baru ke sistem",
                "operationId": "92eb18bc79ced819b38dd9646abcb7fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": ["name", "category", "weight"],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Botol Plastik"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Anorganik"
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data sampah berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trash"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/trash/{id}": {
            "get": {
                "tags": ["Trash"],
                "summary": "Dapatkan detail sampah",
                "description": "Mengambil detail sampah berdasarkan ID",
                "operationId": "021caf20f7f3d390b20e05cf9c49a14c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID sampah",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail sampah berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trash"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sampah tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Trash not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": ["Trash"],
                "summary": "Perbarui data sampah",
                "description": "Memperbarui data sampah yang sudah ada",
                "operationId": "79d159ca25bba3d35932c3336d6a04ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID sampah",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Botol Plastik"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Anorganik"
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data sampah berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trash"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sampah tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Trash not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": ["Trash"],
                "summary": "Hapus data sampah",
                "description": "Menghapus data sampah dari sistem",
                "operationId": "830932ecfc6d248e13e495432c7d0f52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID sampah",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data sampah berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Trash deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sampah tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Trash not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Trash": {
                "title": "Trash",
                "description": "Model untuk data Sampah yang dilacak",
                "properties": {
                    "id": {
                        "description": "ID unik sampah",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Nama/jenis sampah",
                        "type": "string",
                        "example": "Botol Plastik",
                        "maxLength": 255
                    },
                    "category": {
                        "description": "Kategori sampah (organik, anorganik, berbahaya, dll)",
                        "type": "string",
                        "example": "Anorganik",
                        "maxLength": 255
                    },
                    "weight": {
                        "description": "Berat sampah dalam kilogram",
                        "type": "number",
                        "format": "float",
                        "example": 2.5
                    },
                    "created_at": {
                        "description": "Tanggal dan waktu pencatatan sampah",
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-09T10:30:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Tanggal dan waktu pembaruan terakhir",
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-09T10:30:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Gunakan token JWT pada header Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}